Implement a basic Etch-a-Sketch toy
The user draws on the screen by moving their mouse pointer across it.
The number of tiles along each edge of the Etch-a-Sketch screen can be
configured using the controls at the top; however, certain tile counts
produce rendering issues. This is presumably due to floating point
errors when computing the height and width of each tile.